refactor: change conditional export name of define
