+++ src/index.ts
@@ -11,6 +11,6 @@ export { parseArgs, resolveArgs } from 'args-tokens'
 export type { ArgOptions, ArgOptionSchema, ArgValues } from 'args-tokens'
 export * from './cli.ts'
 export { DEFAULT_LOCALE } from './constants.ts'
-export { define } from './define.ts'
+export { define } from './definition.ts'
 export { DefaultTranslation } from './translation.ts'
 export type * from './types.ts'
